# app-making-book-commercial/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: app-making-book API
  version: "1.0"
  description: |
    Backend API for app-making-book. This spec covers the primary
    endpoints; admin sub-routes not listed here follow the same
    Bearer-auth pattern documented on GET /api/admin/dashboard below —
    see backend/src/routes/adminRoutes.ts for the complete list.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        createdAt: { type: string, format: date-time }
...
paths:
...
  /api/auth/register:
    post:
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 10 }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        "409":
          description: Email already registered
  /api/auth/login:
    post:
      summary: Log in — returns tokens, or twoFactorRequired if 2FA is enabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: "Either { accessToken, refreshToken, user } or { twoFactorRequired: true, preAuthToken }"
        "401":
          description: Invalid credentials
        "423":
          description: Account locked (5+ failed attempts)
  /api/auth/2fa/verify-login:
    post:
      summary: Complete login when 2FA is required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [preAuthToken, code]
              properties:
                preAuthToken: { type: string }
                code: { type: string, description: 6-digit TOTP code }
      responses:
        "200": { description: Tokens issued }
        "401": { description: Invalid code or expired pre-auth token }
  /api/auth/refresh:
    post:
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties: { refreshToken: { type: string } }
      responses:
        "200": { description: New access token issued }
        "401": { description: Refresh token invalid, expired, or revoked }
  /api/auth/logout-all:
    post:
      summary: Revoke every access/refresh token on every device
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
  /api/auth/sessions:
    get:
      summary: List active device sessions
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Array of sessions } }
  /api/auth/sessions/{id}:
    delete:
      summary: Revoke a single session (sign out one device)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses: { "200": { description: OK } }
  /api/auth/2fa/setup:
    post:
      summary: Generate a TOTP secret (does not enable 2FA yet)
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: "{ otpauthUrl, secret }" } }
  /api/auth/2fa/enable:
    post:
      summary: Confirm a TOTP code to actually enable 2FA
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { code: { type: string } }
      responses:
        "200": { description: 2FA enabled }
        "401": { description: Invalid code }
  /api/auth/2fa/disable:
    post:
      summary: Turn off 2FA
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
...
  /api/auth/account:
    delete:
      summary: Delete your account (password required)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { password: { type: string } }
      responses:
        "200": { description: Account and all associated data deleted }
        "401": { description: Incorrect password }
  /api/auth/account/export:
    get:
      summary: Export all personal data (GDPR-style), before deletion
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: JSON export of user, progress, logs, events } }
...